requirements.txt
asgiref==3.8.1
boto3==1.35.78
botocore==1.35.78
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
dill==0.3.6
Django==5.1.4
django-storages==1.14.4
djangorestframework==3.15.2
dnslib==0.9.25
dnspython==2.7.0
ecdsa==0.19.0
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
localstack==4.0.3
localstack-core==4.0.3
localstack-ext==4.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moto==5.0.22
packaging==24.2
plux==1.12.1
psutil==6.1.0
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
Pygments==2.18.0
pyotp==2.9.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
requests==2.32.3
responses==0.25.3
rich==13.9.4
rsa==4.9
s3transfer==0.10.4
semver==3.0.2
six==1.17.0
sqlparse==0.5.3
tabulate==0.9.0
tailer==0.4.1
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
xmltodict==0.14.2
